chore: Refactor Gui to meet code quality guidelines
Currently, adding dialog to the MainWindow required MainWindow to
retrieve a response from Tom.java. This leads to Ui needing to store all
messages in an internal buffer before flushing it out in Tom.java.

Adding an interface Ui and an concrete Gui class that interacts with the
MainWindow directly will avoid having multiple levels of abstraction
within a code fragment.

This allows us to meet the code quality standards set out.

This also allows for subsequent CLI implementations of the Ui through
implementation of Ui interface.